Create sample_data folder
