+++ spec/sample_data/github_api.rb
@@ -1,0 +1,287 @@
+def repo_parsed
+  {
+    "id": 393158885,
+    "node_id": "MDEwOlJlcG9zaXRvcnkzOTMxNTg4ODU=",
+    "name": "little_esty_shop_bulk_discounts",
+    "full_name": "Scott-Borecki/little_esty_shop_bulk_discounts",
+    "private": false,
+    "owner": {
+      "login": "Scott-Borecki",
+      "id": 79381792,
+      "node_id": "MDQ6VXNlcjc5MzgxNzky",
+      "avatar_url": "https://avatars.githubusercontent.com/u/79381792?v=4",
+      "gravatar_id": "",
+      "url": "https://api.github.com/users/Scott-Borecki",
+      "html_url": "https://github.com/Scott-Borecki",
+      "followers_url": "https://api.github.com/users/Scott-Borecki/followers",
+      "following_url": "https://api.github.com/users/Scott-Borecki/following{/other_user}",
+      "gists_url": "https://api.github.com/users/Scott-Borecki/gists{/gist_id}",
+      "starred_url": "https://api.github.com/users/Scott-Borecki/starred{/owner}{/repo}",
+      "subscriptions_url": "https://api.github.com/users/Scott-Borecki/subscriptions",
+      "organizations_url": "https://api.github.com/users/Scott-Borecki/orgs",
+      "repos_url": "https://api.github.com/users/Scott-Borecki/repos",
+      "events_url": "https://api.github.com/users/Scott-Borecki/events{/privacy}",
+      "received_events_url": "https://api.github.com/users/Scott-Borecki/received_events",
+      "type": "User",
+      "site_admin": false
+    },
+    "html_url": "https://github.com/Scott-Borecki/little_esty_shop_bulk_discounts",
+    "description": nil,
+    "fork": true,
+    "url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts",
+    "forks_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/forks",
+    "keys_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/keys{/key_id}",
+    "collaborators_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/collaborators{/collaborator}",
+    "teams_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/teams",
+    "hooks_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/hooks",
+    "issue_events_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/issues/events{/number}",
+    "events_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/events",
+    "assignees_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/assignees{/user}",
+    "branches_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/branches{/branch}",
+    "tags_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/tags",
+    "blobs_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/git/blobs{/sha}",
+    "git_tags_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/git/tags{/sha}",
+    "git_refs_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/git/refs{/sha}",
+    "trees_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/git/trees{/sha}",
+    "statuses_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/statuses/{sha}",
+    "languages_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/languages",
+    "stargazers_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/stargazers",
+    "contributors_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/contributors",
+    "subscribers_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/subscribers",
+    "subscription_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/subscription",
+    "commits_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/commits{/sha}",
+    "git_commits_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/git/commits{/sha}",
+    "comments_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/comments{/number}",
+    "issue_comment_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/issues/comments{/number}",
+    "contents_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/contents/{+path}",
+    "compare_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/compare/{base}...{head}",
+    "merges_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/merges",
+    "archive_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/{archive_format}{/ref}",
+    "downloads_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/downloads",
+    "issues_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/issues{/number}",
+    "pulls_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/pulls{/number}",
+    "milestones_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/milestones{/number}",
+    "notifications_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/notifications{?since,all,participating}",
+    "labels_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/labels{/name}",
+    "releases_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/releases{/id}",
+    "deployments_url": "https://api.github.com/repos/Scott-Borecki/little_esty_shop_bulk_discounts/deployments",
+    "created_at": "2021-08-05T20:05:57Z",
+    "updated_at": "2021-08-11T21:03:01Z",
+    "pushed_at": "2021-08-13T17:41:41Z",
+    "git_url": "git://github.com/Scott-Borecki/little_esty_shop_bulk_discounts.git",
+    "ssh_url": "git@github.com:Scott-Borecki/little_esty_shop_bulk_discounts.git",
+    "clone_url": "https://github.com/Scott-Borecki/little_esty_shop_bulk_discounts.git",
+    "svn_url": "https://github.com/Scott-Borecki/little_esty_shop_bulk_discounts",
+    "homepage": "https://shielded-basin-82582.herokuapp.com/",
+    "size": 1579,
+    "stargazers_count": 0,
+    "watchers_count": 0,
+    "language": "Ruby",
+    "has_issues": true,
+    "has_projects": true,
+    "has_downloads": true,
+    "has_wiki": true,
+    "has_pages": false,
+    "forks_count": 0,
+    "mirror_url": nil,
+    "archived": false,
+    "disabled": false,
+    "open_issues_count": 11,
+    "license": nil,
+    "forks": 0,
+    "open_issues": 11,
+    "watchers": 0,
+    "default_branch": "main",
+    "temp_clone_token": nil,
+    "parent": {
+      "id": 329712040,
+      "node_id": "MDEwOlJlcG9zaXRvcnkzMjk3MTIwNDA=",
+      "name": "little_esty_shop_bulk_discounts",
+      "full_name": "turingschool-examples/little_esty_shop_bulk_discounts",
+      "private": false,
+      "owner": {
+        "login": "turingschool-examples",
+        "id": 8518346,
+        "node_id": "MDEyOk9yZ2FuaXphdGlvbjg1MTgzNDY=",
+        "avatar_url": "https://avatars.githubusercontent.com/u/8518346?v=4",
+        "gravatar_id": "",
+        "url": "https://api.github.com/users/turingschool-examples",
+        "html_url": "https://github.com/turingschool-examples",
+        "followers_url": "https://api.github.com/users/turingschool-examples/followers",
+        "following_url": "https://api.github.com/users/turingschool-examples/following{/other_user}",
+        "gists_url": "https://api.github.com/users/turingschool-examples/gists{/gist_id}",
+        "starred_url": "https://api.github.com/users/turingschool-examples/starred{/owner}{/repo}",
+        "subscriptions_url": "https://api.github.com/users/turingschool-examples/subscriptions",
+        "organizations_url": "https://api.github.com/users/turingschool-examples/orgs",
+        "repos_url": "https://api.github.com/users/turingschool-examples/repos",
+        "events_url": "https://api.github.com/users/turingschool-examples/events{/privacy}",
+        "received_events_url": "https://api.github.com/users/turingschool-examples/received_events",
+        "type": "Organization",
+        "site_admin": false
+      },
+      "html_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts",
+      "description": nil,
+      "fork": false,
+      "url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts",
+      "forks_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/forks",
+      "keys_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/keys{/key_id}",
+      "collaborators_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/collaborators{/collaborator}",
+      "teams_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/teams",
+      "hooks_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/hooks",
+      "issue_events_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues/events{/number}",
+      "events_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/events",
+      "assignees_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/assignees{/user}",
+      "branches_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/branches{/branch}",
+      "tags_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/tags",
+      "blobs_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/blobs{/sha}",
+      "git_tags_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/tags{/sha}",
+      "git_refs_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/refs{/sha}",
+      "trees_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/trees{/sha}",
+      "statuses_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/statuses/{sha}",
+      "languages_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/languages",
+      "stargazers_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/stargazers",
+      "contributors_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/contributors",
+      "subscribers_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/subscribers",
+      "subscription_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/subscription",
+      "commits_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/commits{/sha}",
+      "git_commits_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/commits{/sha}",
+      "comments_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/comments{/number}",
+      "issue_comment_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues/comments{/number}",
+      "contents_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/contents/{+path}",
+      "compare_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/compare/{base}...{head}",
+      "merges_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/merges",
+      "archive_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/{archive_format}{/ref}",
+      "downloads_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/downloads",
+      "issues_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues{/number}",
+      "pulls_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/pulls{/number}",
+      "milestones_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/milestones{/number}",
+      "notifications_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/notifications{?since,all,participating}",
+      "labels_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/labels{/name}",
+      "releases_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/releases{/id}",
+      "deployments_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/deployments",
+      "created_at": "2021-01-14T19:18:26Z",
+      "updated_at": "2021-08-05T22:00:29Z",
+      "pushed_at": "2021-08-05T22:00:25Z",
+      "git_url": "git://github.com/turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "ssh_url": "git@github.com:turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "clone_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "svn_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts",
+      "homepage": nil,
+      "size": 1244,
+      "stargazers_count": 0,
+      "watchers_count": 0,
+      "language": "Ruby",
+      "has_issues": true,
+      "has_projects": true,
+      "has_downloads": true,
+      "has_wiki": true,
+      "has_pages": false,
+      "forks_count": 44,
+      "mirror_url": nil,
+      "archived": false,
+      "disabled": false,
+      "open_issues_count": 3,
+      "license": nil,
+      "forks": 44,
+      "open_issues": 3,
+      "watchers": 0,
+      "default_branch": "main"
+    },
+    "source": {
+      "id": 329712040,
+      "node_id": "MDEwOlJlcG9zaXRvcnkzMjk3MTIwNDA=",
+      "name": "little_esty_shop_bulk_discounts",
+      "full_name": "turingschool-examples/little_esty_shop_bulk_discounts",
+      "private": false,
+      "owner": {
+        "login": "turingschool-examples",
+        "id": 8518346,
+        "node_id": "MDEyOk9yZ2FuaXphdGlvbjg1MTgzNDY=",
+        "avatar_url": "https://avatars.githubusercontent.com/u/8518346?v=4",
+        "gravatar_id": "",
+        "url": "https://api.github.com/users/turingschool-examples",
+        "html_url": "https://github.com/turingschool-examples",
+        "followers_url": "https://api.github.com/users/turingschool-examples/followers",
+        "following_url": "https://api.github.com/users/turingschool-examples/following{/other_user}",
+        "gists_url": "https://api.github.com/users/turingschool-examples/gists{/gist_id}",
+        "starred_url": "https://api.github.com/users/turingschool-examples/starred{/owner}{/repo}",
+        "subscriptions_url": "https://api.github.com/users/turingschool-examples/subscriptions",
+        "organizations_url": "https://api.github.com/users/turingschool-examples/orgs",
+        "repos_url": "https://api.github.com/users/turingschool-examples/repos",
+        "events_url": "https://api.github.com/users/turingschool-examples/events{/privacy}",
+        "received_events_url": "https://api.github.com/users/turingschool-examples/received_events",
+        "type": "Organization",
+        "site_admin": false
+      },
+      "html_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts",
+      "description": nil,
+      "fork": false,
+      "url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts",
+      "forks_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/forks",
+      "keys_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/keys{/key_id}",
+      "collaborators_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/collaborators{/collaborator}",
+      "teams_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/teams",
+      "hooks_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/hooks",
+      "issue_events_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues/events{/number}",
+      "events_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/events",
+      "assignees_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/assignees{/user}",
+      "branches_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/branches{/branch}",
+      "tags_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/tags",
+      "blobs_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/blobs{/sha}",
+      "git_tags_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/tags{/sha}",
+      "git_refs_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/refs{/sha}",
+      "trees_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/trees{/sha}",
+      "statuses_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/statuses/{sha}",
+      "languages_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/languages",
+      "stargazers_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/stargazers",
+      "contributors_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/contributors",
+      "subscribers_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/subscribers",
+      "subscription_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/subscription",
+      "commits_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/commits{/sha}",
+      "git_commits_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/git/commits{/sha}",
+      "comments_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/comments{/number}",
+      "issue_comment_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues/comments{/number}",
+      "contents_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/contents/{+path}",
+      "compare_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/compare/{base}...{head}",
+      "merges_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/merges",
+      "archive_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/{archive_format}{/ref}",
+      "downloads_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/downloads",
+      "issues_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/issues{/number}",
+      "pulls_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/pulls{/number}",
+      "milestones_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/milestones{/number}",
+      "notifications_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/notifications{?since,all,participating}",
+      "labels_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/labels{/name}",
+      "releases_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/releases{/id}",
+      "deployments_url": "https://api.github.com/repos/turingschool-examples/little_esty_shop_bulk_discounts/deployments",
+      "created_at": "2021-01-14T19:18:26Z",
+      "updated_at": "2021-08-05T22:00:29Z",
+      "pushed_at": "2021-08-05T22:00:25Z",
+      "git_url": "git://github.com/turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "ssh_url": "git@github.com:turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "clone_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts.git",
+      "svn_url": "https://github.com/turingschool-examples/little_esty_shop_bulk_discounts",
+      "homepage": nil,
+      "size": 1244,
+      "stargazers_count": 0,
+      "watchers_count": 0,
+      "language": "Ruby",
+      "has_issues": true,
+      "has_projects": true,
+      "has_downloads": true,
+      "has_wiki": true,
+      "has_pages": false,
+      "forks_count": 44,
+      "mirror_url": nil,
+      "archived": false,
+      "disabled": false,
+      "open_issues_count": 3,
+      "license": nil,
+      "forks": 44,
+      "open_issues": 3,
+      "watchers": 0,
+      "default_branch": "main"
+    },
+    "network_count": 44,
+    "subscribers_count": 0
+  }
+end
